Implemented second visual debug for BVH traversal
Changed the way ray depth is calculated to count from 1 to max instead of max to 0. Implemented a way to remember intersected, but un visited nodes in the BVH traversal and used variables across main.cpp, render.cpp and bvh.cpp to turn the flags on and off depending on the ray depth in the recursion.